Development checkpoint: connect ports and interfaces
* Add queue workers to poll port queues, perform interface
  lookups, and forward packets to matching interfaces.
* Add a link status change callback to log link status changes
  for all ports.
* Add a basic 'switch' object to maintain port -> interface
  mappings.
* Add/remove MAC addresses from port on interface creation/deletion.
* Add a netif wrapper object to clean up copy/move semantics for
  the real net interface object.
* Fix up various other minor issues.

+++ src/packetio/stack/dpdk/lwip.h
@@ -59,7 +59,7 @@ public:
 private:
     driver::generic_driver& m_driver;
     std::unordered_map<size_t, std::unique_ptr<net_interface>> m_interfaces;
-    int m_idx;
+    size_t m_idx;
 };
 
 }
